+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+#### v0.0.4
+- Fixed relative import of geopandas.
+- Added platform profiles export tool.
+- Minor general improvements.
+
 #### v0.0.3
 - Added readme to the overall tool.
 - Improved profile metadata parsing